- ещё более глубокая структуризация по папкам - созданы viewModel для всех функциональных экранов - handler заменен на корутины
