--- app/src/main/java/com/example/playlistmaker/App.kt
@@ -1,43 +1,0 @@
-package com.example.playlistmaker
-
-import android.app.Application
-import android.content.Context
-import androidx.appcompat.app.AppCompatDelegate
-
-class App : Application() {
-
-    var darkTheme: Boolean = false
-
-    override fun onCreate() {
-        super.onCreate()
-
-        val sharedPreferences = getSharedPreferences("AppPreferences", Context.MODE_PRIVATE)
-        darkTheme = sharedPreferences.getBoolean("darkTheme", false)
-
-        AppCompatDelegate.setDefaultNightMode(
-            if (darkTheme) {
-                AppCompatDelegate.MODE_NIGHT_YES
-            } else {
-                AppCompatDelegate.MODE_NIGHT_NO
-            }
-        )
-    }
-
-    fun switchTheme(darkThemeEnabled: Boolean) {
-        darkTheme = darkThemeEnabled
-
-        val sharedPreferences = getSharedPreferences("AppPreferences", Context.MODE_PRIVATE)
-        with(sharedPreferences.edit()) {
-            putBoolean("darkTheme", darkThemeEnabled)
-            apply()
-        }
-
-        AppCompatDelegate.setDefaultNightMode(
-            if (darkThemeEnabled) {
-                AppCompatDelegate.MODE_NIGHT_YES
-            } else {
-                AppCompatDelegate.MODE_NIGHT_NO
-            }
-        )
-    }
-}
