--- app/src/main/java/com/example/playlistmaker/data/network/RetrofitInstanse.kt
@@ -1,23 +1,0 @@
-package com.example.playlistmaker.data.network
-
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitInstance {
-    val api: iTunesAPI by lazy {
-        val logging = HttpLoggingInterceptor().apply {
-            level = HttpLoggingInterceptor.Level.BODY
-        }
-        val client = OkHttpClient.Builder()
-            .addInterceptor(logging)
-            .build()
-        Retrofit.Builder()
-            .baseUrl("https://itunes.apple.com")
-            .client(client)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(iTunesAPI::class.java)
-    }
-}
