+++ app/src/main/java/com/example/playlistmaker/data/search/TrackRepositoryImpl.kt
@@ -1,0 +1,39 @@
+package com.example.playlistmaker.data.search
+
+import com.example.playlistmaker.data.NetworkClient
+import com.example.playlistmaker.data.dto.TrackSearchRequest
+import com.example.playlistmaker.data.dto.TrackSearchResponse
+import com.example.playlistmaker.domain.api.search.TrackRepository
+import com.example.playlistmaker.domain.model.track.Track
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class TrackRepositoryImpl(private val networkClient: NetworkClient) : TrackRepository {
+
+    override fun searchTrack(expression: String): List<Track> {
+        val response = networkClient.doRequest(TrackSearchRequest(expression))
+        if (response.resultCode == 200) {
+            return (response as TrackSearchResponse).results.map {
+                Track(
+                    it.trackId,
+                    it.trackName,
+                    it.artistName,
+                    formatTrackTime(it.trackTimeMillis),
+                    it.artworkUrl100,
+                    it.collectionName,
+                    it.releaseDate,
+                    it.primaryGenreName,
+                    it.country,
+                    it.previewUrl
+                )
+            }
+        } else {
+            return emptyList()
+        }
+    }
+
+    private fun formatTrackTime(trackTimeMillis: Long): String {
+        return SimpleDateFormat("mm:ss", Locale.getDefault()).format(trackTimeMillis)
+    }
+    }
+
