+++ app/src/main/java/com/example/playlistmaker/data/sharing/ExternalNavigator.kt
@@ -1,0 +1,39 @@
+package com.example.playlistmaker.data.sharing
+
+import android.content.Context
+import android.content.Intent
+import android.net.Uri
+import com.example.playlistmaker.domain.model.sharing.EmailData
+
+class ExternalNavigator(private val context: Context) {
+
+    fun shareLink(link: String) {
+        val shareIntent = Intent(Intent.ACTION_SEND).apply {
+            type = "text/plain"
+            putExtra(Intent.EXTRA_TEXT, link)
+            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+        }
+        val chooserIntent = Intent.createChooser(shareIntent, "Поделиться").apply {
+            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+        }
+        context.startActivity(chooserIntent)
+    }
+
+    fun openLink(link: String) {
+        val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(link)).apply {
+            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+        }
+        context.startActivity(browserIntent)
+    }
+
+    fun openEmail(emailData: EmailData) {
+        val emailIntent = Intent(Intent.ACTION_SENDTO).apply {
+            data = Uri.parse("mailto:")
+            putExtra(Intent.EXTRA_EMAIL, arrayOf(emailData.recipient))
+            putExtra(Intent.EXTRA_SUBJECT, emailData.subject)
+            putExtra(Intent.EXTRA_TEXT, emailData.body)
+            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+        }
+        context.startActivity(emailIntent)
+    }
+}
