+++ app/src/main/java/com/example/playlistmaker/domain/api/search/TrackRepository.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.domain.api.search
+
+import com.example.playlistmaker.domain.model.track.Track
+
+interface TrackRepository {
+    fun searchTrack(expression: String): List<Track>
+}
