+++ app/src/main/java/com/example/playlistmaker/domain/api/searchHistory/SearchHistoryInteractor.kt
@@ -1,6 +1,6 @@
 package com.example.playlistmaker.domain.api.searchHistory
 
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.domain.model.track.Track
 
 interface SearchHistoryInteractor {
     fun saveTrack(track: Track)
