+++ app/src/main/java/com/example/playlistmaker/domain/api/searchHistory/SearchHistoryRepository.kt
@@ -1,6 +1,6 @@
 package com.example.playlistmaker.domain.api.searchHistory
 
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.domain.model.track.Track
 
 interface SearchHistoryRepository {
     fun saveTrack(track: Track)
