+++ app/src/main/java/com/example/playlistmaker/domain/api/sharing/SharingInteractor.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.domain.api.sharing
+
+interface SharingInteractor {
+    fun shareApp()
+    fun openTerms()
+    fun openSupport()
+}
