--- app/src/main/java/com/example/playlistmaker/domain/api/track/TrackRepository.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.domain.api.track
-
-import com.example.playlistmaker.domain.models.Track
-
-interface TrackRepository {
-    fun searchTrack(expression: String): List<Track>
-}
