+++ app/src/main/java/com/example/playlistmaker/domain/impl/settings/SettingsInteractorImpl.kt
@@ -1,11 +1,12 @@
 package com.example.playlistmaker.domain.impl.settings
 
-import com.example.playlistmaker.App
+import com.example.playlistmaker.app.App
 import com.example.playlistmaker.domain.api.settings.SettingsInteractor
 
 class SettingsInteractorImpl(private val app: App) : SettingsInteractor {
+
     override var isDarkTheme: Boolean
-        get() = app.darkTheme
+        get() = app.isDarkTheme()
         set(value) { app.switchTheme(value) }
 
     override fun switchTheme(isDark: Boolean) = app.switchTheme(isDark)
