--- app/src/main/java/com/example/playlistmaker/domain/impl/track/TrackInteractorImpl.kt
@@ -1,17 +1,0 @@
-package com.example.playlistmaker.domain.impl.track
-
-import com.example.playlistmaker.domain.api.track.TrackInteractor
-import com.example.playlistmaker.domain.api.track.TrackRepository
-import java.util.concurrent.Executors
-
-class TrackInteractorImpl(private val repository: TrackRepository) : TrackInteractor {
-
-    private val executor = Executors.newCachedThreadPool()
-
-    override fun searchTrack(expression: String, consumer: TrackInteractor.TrackConsumer) {
-        executor.execute {
-            val foundTracks = repository.searchTrack(expression)
-            consumer.consume(foundTracks)
-        }
-    }
-}
