+++ app/src/main/java/com/example/playlistmaker/domain/model/sharing/EmailData.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.domain.model.sharing
+
+data class EmailData(
+    val recipient: String,
+    val subject: String,
+    val body: String
+)
