+++ app/src/main/java/com/example/playlistmaker/domain/model/track/Track.kt
@@ -1,0 +1,16 @@
+package com.example.playlistmaker.domain.model.track
+
+data class Track(
+    val trackId: Int,
+    val trackName: String,
+    val artistName: String,
+    val trackTimeMillis: String,
+    val artworkUrl100: String,
+    val collectionName: String,
+    val releaseDate: String,
+    val primaryGenreName: String,
+    val country: String,
+    val previewUrl: String
+)
+
+
