+++ app/src/main/java/com/example/playlistmaker/ui/mediateka/MediatekaActivity.kt
@@ -1,0 +1,26 @@
+package com.example.playlistmaker.ui.mediateka
+
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import com.example.playlistmaker.R
+import com.example.playlistmaker.databinding.ActivityMediatekaBinding
+
+class MediatekaActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMediatekaBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        binding = ActivityMediatekaBinding.inflate(layoutInflater)
+        val view = binding.root
+        setContentView(view)
+
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }
+    }
+}
