+++ app/src/main/java/com/example/playlistmaker/ui/search/view_model/SearchViewModelFactory.kt
@@ -1,0 +1,19 @@
+package com.example.playlistmaker.ui.search.view_model
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.example.playlistmaker.domain.api.search.TrackInteractor
+import com.example.playlistmaker.domain.api.searchHistory.SearchHistoryInteractor
+
+class SearchViewModelFactory(
+    private val trackInteractor: TrackInteractor,
+    private val searchHistoryInteractor: SearchHistoryInteractor
+) : ViewModelProvider.Factory {
+    @Suppress("UNCHECKED_CAST")
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(SearchViewModel::class.java)) {
+            return SearchViewModel(trackInteractor, searchHistoryInteractor) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
