+++ app/src/main/java/com/example/playlistmaker/ui/settings/activity/SettingsActivity.kt
@@ -1,0 +1,58 @@
+package com.example.playlistmaker.ui.settings.activity
+
+import android.os.Bundle
+import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import com.example.playlistmaker.app.App
+import com.example.playlistmaker.R
+import com.example.playlistmaker.databinding.ActivitySettingsBinding
+import com.example.playlistmaker.ui.settings.view_model.SettingsViewModel
+import com.example.playlistmaker.ui.settings.view_model.SettingsViewModelFactory
+
+class SettingsActivity : AppCompatActivity() {
+    private lateinit var binding: ActivitySettingsBinding
+    private val viewModel: SettingsViewModel by viewModels {
+        SettingsViewModelFactory(
+            (applicationContext as App).provideSharingInteractor(),
+            (applicationContext as App).provideSettingsInteractor()
+        )
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivitySettingsBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }
+
+        binding.backFromSettings.setOnClickListener {
+            finish()
+        }
+
+        binding.themeSwitcher.setOnCheckedChangeListener { _, isChecked ->
+            viewModel.switchTheme(isChecked)
+        }
+
+        binding.shareButton.setOnClickListener {
+            viewModel.shareApp()
+        }
+
+        binding.supportButton.setOnClickListener {
+            viewModel.openSupport()
+        }
+
+        binding.agreementButton.setOnClickListener {
+            viewModel.openTerms()
+        }
+
+        viewModel.isDarkTheme.observe(this) { isDark ->
+            binding.themeSwitcher.isChecked = isDark
+        }
+    }
+}
