+++ app/src/main/java/com/example/playlistmaker/ui/settings/view_model/SettingsViewModel.kt
@@ -1,0 +1,37 @@
+package com.example.playlistmaker.ui.settings.view_model
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.example.playlistmaker.domain.api.settings.SettingsInteractor
+import com.example.playlistmaker.domain.api.sharing.SharingInteractor
+
+class SettingsViewModel(
+    private val sharingInteractor: SharingInteractor,
+    private val settingsInteractor: SettingsInteractor
+) : ViewModel() {
+
+    private val _isDarkTheme = MutableLiveData<Boolean>()
+    val isDarkTheme: LiveData<Boolean> = _isDarkTheme
+
+    init {
+        _isDarkTheme.value = settingsInteractor.isDarkTheme
+    }
+
+    fun switchTheme(isDark: Boolean) {
+        settingsInteractor.switchTheme(isDark)
+        _isDarkTheme.value = isDark
+    }
+
+    fun shareApp() {
+        sharingInteractor.shareApp()
+    }
+
+    fun openSupport() {
+        sharingInteractor.openSupport()
+    }
+
+    fun openTerms() {
+        sharingInteractor.openTerms()
+    }
+}
