+++ app/src/main/java/com/example/playlistmaker/ui/settings/view_model/SettingsViewModelFactory.kt
@@ -1,0 +1,21 @@
+package com.example.playlistmaker.ui.settings.view_model
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.example.playlistmaker.domain.api.settings.SettingsInteractor
+import com.example.playlistmaker.domain.api.sharing.SharingInteractor
+
+class SettingsViewModelFactory(
+    private val sharingInteractor: SharingInteractor,
+    private val settingsInteractor: SettingsInteractor
+) : ViewModelProvider.Factory {
+    @Suppress("UNCHECKED_CAST")
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(SettingsViewModel::class.java)) {
+            return SettingsViewModel(sharingInteractor, settingsInteractor) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
+
+
+
