+++ app/src/main/java/com/example/playlistmaker/ui/track/TrackViewHolder.kt
@@ -1,0 +1,38 @@
+package com.example.playlistmaker.ui.track
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+import com.example.playlistmaker.R
+import com.example.playlistmaker.domain.model.track.Track
+
+class TrackViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    private val trackName: TextView = itemView.findViewById(R.id.trackTitle)
+    private val artistName: TextView = itemView.findViewById(R.id.trackArtist)
+    private val trackTime: TextView = itemView.findViewById(R.id.trackLength)
+    private val artwork: ImageView = itemView.findViewById(R.id.trackCover)
+
+    fun bind(track: Track) {
+        trackName.text = track.trackName
+        artistName.text = track.artistName
+        trackTime.text = track.trackTimeMillis
+        Glide.with(itemView)
+            .load(track.artworkUrl100)
+            .centerCrop()
+            .transform(RoundedCorners(2))
+            .placeholder(R.drawable.placeholder)
+            .into(artwork)
+    }
+    companion object {
+        fun create(parent: ViewGroup): TrackViewHolder {
+            val view = LayoutInflater.from(parent.context)
+                .inflate(R.layout.search_track, parent, false)
+            return TrackViewHolder(view)
+        }
+    }
+}
