+++ app/src/main/res/layout/activity_audioplayer.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/main"
-    tools:context=".presentation.ui.audioplayer.AudioPlayerActivity"
+    tools:context=".ui.audioplayer.activity.AudioPlayerActivity"
     style="@style/SettingsBackground">
 
 <androidx.constraintlayout.widget.ConstraintLayout
