+++ app/src/main/res/layout/activity_mediateka.xml
@@ -5,6 +5,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.mediateka.MediatekaActivity">
+    tools:context=".ui.mediateka.MediatekaActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
